movie group and ungroup methods from abstract
