+++ jhotdraw-core/src/main/java/org/jhotdraw/draw/action/GroupAction.java
@@ -8,14 +8,12 @@
 package org.jhotdraw.draw.action;
 
 import dk.sdu.mmmi.featuretracer.lib.FeatureEntryPoint;
+import org.jhotdraw.draw.DrawingEditor;
 import org.jhotdraw.draw.figure.Figure;
 import org.jhotdraw.draw.figure.CompositeFigure;
 import org.jhotdraw.draw.figure.GroupFigure;
 
 import java.util.*;
-
-import org.jhotdraw.draw.*;
-import org.jhotdraw.util.ResourceBundleUtil;
 
 /**
  * GroupAction.
@@ -23,23 +21,23 @@ import org.jhotdraw.util.ResourceBundleUtil;
  * @author Werner Randelshofer
  * @version $Id$
  */
+
 public class GroupAction extends AbstractGroupingAction {
 
     private static final long serialVersionUID = 1L;
     public static final String ID = "edit.groupSelection";
+    private final UngroupAction ungroupAction;
 
     public GroupAction(DrawingEditor editor) {
         this(editor, new GroupFigure());
     }
 
-
-
     public GroupAction(DrawingEditor editor, CompositeFigure compositeFigure) {
         super(editor);
         this.compositeFigure = compositeFigure;
-        ResourceBundleUtil labels = ResourceBundleUtil.getBundle("org.jhotdraw.draw.Labels");
         labels.configureAction(this, ID);
         updateEnabledState();
+        ungroupAction = new UngroupAction(editor, compositeFigure);
     }
 
     protected boolean canGroup() {
@@ -67,9 +65,34 @@ public class GroupAction extends AbstractGroupingAction {
         }
     }
 
+    public void groupFigures(CompositeFigure group, Collection<Figure> figures) {
+
+        // assertions
+        assert group != null;
+        assert figures.size() > 1;
+
+        // returns a copy of the figures sorted from back to front
+        Collection<Figure> sorted = view.getDrawing().sort(figures);
+        int index = view.getDrawing().indexOf(sorted.iterator().next());
+        view.getDrawing().basicRemoveAll(figures);
+        view.clearSelection();
+
+        // adding the composite figure to the drawing
+        view.getDrawing().add(index, group);
+        group.willChange();
+
+        // Adding figures to group
+        for (Figure f : sorted) {
+            f.willChange();
+            group.basicAdd(f);
+        }
+        group.changed();
+        view.addToSelection(group);
+    }
+
     @Override
     void undoAction(Collection<Figure> figures, CompositeFigure group) {
-        ungroupFigures(group);
+        ungroupAction.ungroupFigures(group);
     }
 
     @Override
